Init for locking checkoutComplete
