+++ saleor/checkout/models.py
@@ -96,6 +96,9 @@ class Checkout(ModelWithMetadata):
     voucher_code = models.CharField(max_length=12, blank=True, null=True)
     gift_cards = models.ManyToManyField(GiftCard, blank=True, related_name="checkouts")
 
+    redirect_url = models.URLField(blank=True, null=True)
+    tracking_code = models.CharField(max_length=255, blank=True, null=True)
+
     objects = CheckoutQueryset.as_manager()
 
     class Meta:
