+++ saleor/graphql/checkout/mutations.py
@@ -1,12 +1,14 @@
-from typing import List, Optional, Tuple
+from typing import List, Optional, Tuple, Union
 
 import graphene
 from django.conf import settings
+from django.contrib.auth.models import AnonymousUser
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import transaction
 from django.db.models import Prefetch
 
 from ...account.error_codes import AccountErrorCode
+from ...account.models import User
 from ...checkout import models
 from ...checkout.calculations import calculate_checkout_total_with_gift_cards
 from ...checkout.error_codes import CheckoutErrorCode
@@ -17,6 +19,7 @@ from ...checkout.utils import (
     change_billing_address_in_checkout,
     change_shipping_address_in_checkout,
     create_order,
+    get_order,
     get_user_checkout,
     get_valid_shipping_methods_for_checkout,
     prepare_order_data,
@@ -28,10 +31,10 @@ from ...core.exceptions import InsufficientStock, PermissionDenied, ProductNotPu
 from ...core.permissions import OrderPermissions
 from ...core.taxes import TaxError
 from ...core.utils.url import validate_storefront_url
-from ...discount import models as voucher_model
+from ...discount import DiscountInfo, models as voucher_model
 from ...graphql.checkout.utils import clean_checkout_payment, clean_checkout_shipping
+from ...order import models as order_models
 from ...payment import PaymentError, gateway, models as payment_models
-from ...payment.interface import AddressData
 from ...payment.utils import store_customer_id
 from ...product import models as product_models
 from ...warehouse.availability import check_stock_quantity, get_available_quantity
@@ -789,28 +792,11 @@ class CheckoutComplete(BaseMutation):
             )
 
     @classmethod
-    def perform_mutation(cls, _root, info, checkout_id, store_source, **data):
-        checkout = cls.get_node_or_error(
-            info,
-            checkout_id,
-            only_type=Checkout,
-            field="checkout_id",
-            qs=models.Checkout.objects.prefetch_related(
-                "gift_cards",
-                "lines",
-                Prefetch(
-                    "payments",
-                    queryset=payment_models.Payment.objects.prefetch_related(
-                        "order", "order__lines"
-                    ),
-                ),
-            ).select_related("shipping_method", "shipping_method__shipping_zone"),
-        )
+    @transaction.atomic
+    def _prepare_checkout(cls, checkout: models.Checkout, context, redirect_url):
         lines = list(checkout)
 
-        discounts = info.context.discounts
-        user = info.context.user
-
+        discounts = context.discounts
         clean_checkout_shipping(checkout, lines, discounts, CheckoutErrorCode)
         clean_checkout_payment(checkout, lines, discounts, CheckoutErrorCode)
 
@@ -818,80 +804,144 @@ class CheckoutComplete(BaseMutation):
 
         cls.validate_payment_amount(discounts, payment, checkout)
 
-        redirect_url = data.get("redirect_url", "")
         if redirect_url:
             try:
                 validate_storefront_url(redirect_url)
             except ValidationError as error:
                 raise ValidationError(
-                    {"redirect_url": error}, code=AccountErrorCode.INVALID
+                    {"redirect_url": error}, code=AccountErrorCode.INVALID.value
                 )
 
-        with transaction.atomic():
-            try:
+        to_update = []
+        if redirect_url and redirect_url != checkout.redirect_url:
+            checkout.redirect_url = redirect_url
+            to_update.append("redirect_url")
+
+        tracking_code = analytics.get_client_id(context)
+        if tracking_code and tracking_code != checkout.tracking_code:
+            checkout.tracking_code = tracking_code
+            to_update.append("tracking_code")
+
+        if to_update:
+            checkout.save()
+
+    @classmethod
+    @transaction.atomic
+    def _convert_checkout_to_order(
+        cls,
+        checkout: models.Checkout,
+        order_data: dict,
+        user: Union[User, AnonymousUser],
+    ) -> order_models.Order:
+        order = create_order(
+            checkout=checkout, order_data=order_data, user=user  # type: ignore
+        )
+        # remove checkout after order is successfully created
+        checkout.delete()
+        return order
+
+    @classmethod
+    def _get_order_data(
+        cls, checkout: models.Checkout, discounts: List[DiscountInfo]
+    ) -> dict:
+        try:
+            with transaction.atomic():
                 order_data = prepare_order_data(
-                    checkout=checkout,
-                    lines=lines,
-                    tracking_code=analytics.get_client_id(info.context),
-                    discounts=discounts,
+                    checkout=checkout, lines=list(checkout), discounts=discounts,
                 )
-            except InsufficientStock as e:
-                gateway.payment_refund_or_void(payment)
-                raise ValidationError(
-                    f"Insufficient product stock: {e.item}", code=e.code
-                )
-            except voucher_model.NotApplicable:
-                raise ValidationError(
-                    "Voucher not applicable",
-                    code=CheckoutErrorCode.VOUCHER_NOT_APPLICABLE,
-                )
-            except TaxError as tax_error:
-                raise ValidationError(
-                    "Unable to calculate taxes - %s" % str(tax_error),
-                    code=CheckoutErrorCode.TAX_ERROR,
-                )
+        except InsufficientStock as e:
+            raise ValidationError(f"Insufficient product stock: {e.item}", code=e.code)
+        except voucher_model.NotApplicable:
+            raise ValidationError(
+                "Voucher not applicable",
+                code=CheckoutErrorCode.VOUCHER_NOT_APPLICABLE.value,
+            )
+        except TaxError as tax_error:
+            raise ValidationError(
+                "Unable to calculate taxes - %s" % str(tax_error),
+                code=CheckoutErrorCode.TAX_ERROR.value,
+            )
+        return order_data
 
-        billing_address = order_data["billing_address"]
-        shipping_address = order_data.get("shipping_address", None)
-
-        billing_address = AddressData(**billing_address.as_data())
-
-        if shipping_address is not None:
-            shipping_address = AddressData(**shipping_address.as_data())
-
+    @classmethod
+    def _process_payment(
+        cls,
+        payment: payment_models.Payment,
+        store_source: bool,
+        payment_data: Optional[dict],
+        order_data: dict,
+    ) -> payment_models.Transaction:
         payment_confirmation = payment.to_confirm
         try:
             if payment_confirmation:
-                txn = gateway.confirm(payment, additional_data=data.get("payment_data"))
+                txn = gateway.confirm(payment, additional_data=payment_data)
             else:
                 txn = gateway.process_payment(
                     payment=payment,
                     token=payment.token,
                     store_source=store_source,
-                    additional_data=data.get("payment_data"),
+                    additional_data=payment_data,
                 )
             if not txn.is_success:
                 raise PaymentError(txn.error)
-
         except PaymentError as e:
             abort_order_data(order_data)
-            raise ValidationError(str(e), code=CheckoutErrorCode.PAYMENT_ERROR)
+            raise ValidationError(str(e), code=CheckoutErrorCode.PAYMENT_ERROR.value)
+        return txn
 
+    @classmethod
+    def complete_checkout(cls, checkout: models.Checkout, context, data, store_source):
+        order = get_order(checkout.token)
+        if order:
+            return CheckoutComplete(
+                order=order, confirmation_needed=False, confirmation_data={},
+            )
+        cls._prepare_checkout(
+            checkout=checkout,
+            context=context,
+            redirect_url=data.get("redirect_url", ""),
+        )
+
+        payment = checkout.get_last_active_payment()  # type: ignore
+
+        try:
+            order_data = cls._get_order_data(checkout, context.discounts)
+        except ValidationError as error:
+            gateway.payment_refund_or_void(payment)
+            raise error
+
+        txn = cls._process_payment(
+            payment=payment,  # type: ignore
+            store_source=store_source,
+            payment_data=data.get("payment_data"),
+            order_data=order_data,
+        )
+
+        user = context.user
         if txn.customer_id and user.is_authenticated:
-            store_customer_id(user, payment.gateway, txn.customer_id)
-
-        order = None
+            store_customer_id(user, payment.gateway, txn.customer_id)  # type: ignore
 
         if not txn.action_required:
-            order = create_order(
-                checkout=checkout,
-                order_data=order_data,
-                user=user,
-                redirect_url=redirect_url,
-            )
+            try:
+                with transaction.atomic():
+                    order = get_order(checkout.token)
+                    if order:
+                        # Order was created asynchronously, we can release the lock made
+                        # on order_data
+                        abort_order_data(order_data)
+                    else:
+                        order = cls._convert_checkout_to_order(
+                            checkout=checkout, order_data=order_data, user=user
+                        )
+            except InsufficientStock as e:
+                abort_order_data(order_data)
+                gateway.payment_refund_or_void(payment)
+                raise ValidationError(
+                    f"Insufficient product stock: {e.item}", code=e.code
+                )
 
-            # remove checkout after order is successfully paid
-            checkout.delete()
+        if not order:
+            abort_order_data(order_data)
 
         # If gateway returns information that additional steps are required we need
         # to inform the frontend and pass all required data
@@ -900,6 +950,40 @@ class CheckoutComplete(BaseMutation):
             confirmation_needed=txn.action_required,
             confirmation_data=txn.action_required_data if txn.action_required else {},
         )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, checkout_id, store_source, **data):
+        checkout_token = from_global_id_strict_type(
+            checkout_id, Checkout, field="checkout_id"
+        )
+        order = get_order(checkout_token)
+        if order:
+            # The order is already created. We return it as a success checkoutComplete
+            # response. Order is anonymized for not logged in user
+            return CheckoutComplete(
+                order=order, confirmation_needed=False, confirmation_data={}
+            )
+
+        with transaction.atomic():
+            checkout = cls.get_node_or_error(
+                info,
+                checkout_id,
+                only_type=Checkout,
+                field="checkout_id",
+                qs=models.Checkout.objects.select_for_update(of=("self",))
+                .prefetch_related(
+                    "gift_cards",
+                    "lines",
+                    Prefetch(
+                        "payments",
+                        queryset=payment_models.Payment.objects.prefetch_related(
+                            "order", "order__lines"
+                        ),
+                    ),
+                )
+                .select_related("shipping_method", "shipping_method__shipping_zone"),
+            )
+        return cls.complete_checkout(checkout, info.context, data, store_source)
 
 
 class CheckoutAddPromoCode(BaseMutation):
