+++ saleor/graphql/checkout/utils.py
@@ -1,4 +1,4 @@
-from typing import Iterable, Union
+from typing import Iterable, Type, Union
 
 from django.core.exceptions import ValidationError
 
@@ -13,7 +13,7 @@ def clean_checkout_shipping(
     checkout: Checkout,
     lines: Iterable[CheckoutLine],
     discounts: Iterable[DiscountInfo],
-    error_code: Union[CheckoutErrorCode, PaymentErrorCode],
+    error_code: Union[Type[CheckoutErrorCode], Type[PaymentErrorCode]],
 ):
     if checkout.is_shipping_required():
         if not checkout.shipping_method:
@@ -21,7 +21,7 @@ def clean_checkout_shipping(
                 {
                     "shipping_method": ValidationError(
                         "Shipping method is not set",
-                        code=error_code.SHIPPING_METHOD_NOT_SET,
+                        code=error_code.SHIPPING_METHOD_NOT_SET.value,
                     )
                 }
             )
@@ -30,7 +30,7 @@ def clean_checkout_shipping(
                 {
                     "shipping_address": ValidationError(
                         "Shipping address is not set",
-                        code=error_code.SHIPPING_ADDRESS_NOT_SET,
+                        code=error_code.SHIPPING_ADDRESS_NOT_SET.value,
                     )
                 }
             )
@@ -39,21 +39,22 @@ def clean_checkout_shipping(
                 {
                     "shipping_method": ValidationError(
                         "Shipping method is not valid for your shipping address",
-                        code=error_code.INVALID_SHIPPING_METHOD,
+                        code=error_code.INVALID_SHIPPING_METHOD.value,
                     )
                 }
             )
 
 
 def clean_billing_address(
-    checkout: Checkout, error_code: Union[CheckoutErrorCode, PaymentErrorCode],
+    checkout: Checkout,
+    error_code: Union[Type[CheckoutErrorCode], Type[PaymentErrorCode]],
 ):
     if not checkout.billing_address:
         raise ValidationError(
             {
                 "billing_address": ValidationError(
                     "Billing address is not set",
-                    code=error_code.BILLING_ADDRESS_NOT_SET,
+                    code=error_code.BILLING_ADDRESS_NOT_SET.value,
                 )
             }
         )
@@ -63,11 +64,11 @@ def clean_checkout_payment(
     checkout: Checkout,
     lines: Iterable[CheckoutLine],
     discounts: Iterable[DiscountInfo],
-    error_code: CheckoutErrorCode,
+    error_code: Type[CheckoutErrorCode],
 ):
     clean_billing_address(checkout, error_code)
     if not is_fully_paid(checkout, lines, discounts):
         raise ValidationError(
             "Provided payment methods can not cover the checkout's total amount",
-            code=error_code.CHECKOUT_NOT_FULLY_PAID,
+            code=error_code.CHECKOUT_NOT_FULLY_PAID.value,
         )
