+++ saleor/payment/gateways/adyen/tests/webhooks/cassettes/test_handle_authorization_for_checkout_wrong_total.yaml
@@ -1,0 +1,46 @@
+interactions:
+- request:
+    body: '{"merchantAccount": "SaleorECOM", "originalReference": "852595499936560C",
+      "reference": "UGF5bWVudDoxMDg=", "applicationInfo": {"adyenLibrary": {"name":
+      "adyen-python-api-library", "version": "3.0.0"}}}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '202'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - adyen-python-api-library/3.0.0
+      x-api-key:
+      - test_key
+    method: POST
+    uri: https://pal-test.adyen.com/pal/servlet/Payment/v49/cancel
+  response:
+    body:
+      string: '{"status":401,"errorCode":"000","message":"HTTP Status Response - Unauthorized","errorType":"security"}'
+    headers:
+      Connection:
+      - Keep-Alive
+      Content-Type:
+      - application/json
+      Date:
+      - Tue, 18 Aug 2020 13:28:46 GMT
+      Keep-Alive:
+      - timeout=15, max=100
+      Server:
+      - Apache
+      Set-Cookie:
+      - JSESSIONID=3ADD2912A60201D4D654B56E5610FA80.test103e; Path=/pal; Secure; HttpOnly
+      Transfer-Encoding:
+      - chunked
+      WWW-Authenticate:
+      - BASIC realm="Adyen PAL Service Authentication"
+    status:
+      code: 401
+      message: '401'
+version: 1
