+++ saleor/payment/gateways/adyen/webhooks.py
@@ -11,6 +11,7 @@ import graphene
 from django.contrib.auth.hashers import check_password
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.handlers.wsgi import WSGIRequest
+from django.db import transaction
 from django.http import (
     HttpResponse,
     HttpResponseBadRequest,
@@ -21,6 +22,7 @@ from django.http.request import HttpHeaders
 from django.shortcuts import redirect
 from graphql_relay import from_global_id
 
+from ....checkout.models import Checkout
 from ....core.utils.url import prepare_url
 from ....order.actions import (
     cancel_order,
@@ -31,10 +33,15 @@ from ....order.actions import (
 from ....order.events import external_notification_event
 from ....payment.models import Payment, Transaction
 from ... import ChargeStatus, PaymentError, TransactionKind
-from ...gateway import capture
+from ...gateway import capture, payment_refund_or_void
 from ...interface import GatewayConfig, GatewayResponse
 from ...utils import create_payment_information, create_transaction, gateway_postprocess
-from .utils import FAILED_STATUSES, api_call, from_adyen_price
+from .utils import (
+    FAILED_STATUSES,
+    api_call,
+    create_order_from_checkout,
+    from_adyen_price,
+)
 
 
 def get_payment(payment_id: Optional[str]) -> Optional[Payment]:
@@ -45,12 +52,25 @@ def get_payment(payment_id: Optional[str]) -> Optional[Payment]:
     except UnicodeDecodeError:
         return None
     payment = (
-        Payment.objects.prefetch_related("order")
-        .select_related()
+        Payment.objects.prefetch_related("order", "checkout")
+        .select_for_update(of=("self",))
         .filter(id=db_payment_id)
         .first()
     )
     return payment
+
+
+def get_checkout(payment: Payment) -> Optional[Checkout]:
+    if not payment.checkout:
+        return None
+    # Lock checkout in the same way as in checkoutComplete
+    return (
+        Checkout.objects.select_for_update(of=("self",))
+        .prefetch_related("gift_cards", "lines",)
+        .select_related("shipping_method", "shipping_method__shipping_zone")
+        .filter(token=payment.checkout.pk)
+        .first()
+    )
 
 
 def get_transaction(
@@ -101,11 +121,13 @@ def create_payment_notification_for_order(
     )
 
 
+@transaction.atomic
 def handle_authorization(notification: Dict[str, Any], gateway_config: GatewayConfig):
     payment = get_payment(notification.get("merchantReference"))
     if not payment:
         # We don't know anything about that payment
         return
+    checkout = get_checkout(payment)
     if payment.charge_status in {
         ChargeStatus.FULLY_CHARGED,
         ChargeStatus.PARTIALLY_CHARGED,
@@ -132,6 +154,15 @@ def handle_authorization(notification: Dict[str, Any], gateway_config: GatewayCo
     transaction = create_new_transaction(notification, payment, kind)
     reason = notification.get("reason", "-")
 
+    if not payment.order and transaction.is_success and checkout:
+        try:
+            order = create_order_from_checkout(checkout)
+        except Exception:
+            payment_refund_or_void(payment)
+            return
+        checkout.delete()
+        payment.order = order
+
     success_msg = f"Adyen: The payment  {transaction_id} request  was successful."
     failed_msg = f"Adyen: The payment {transaction_id} request failed. Reason: {reason}"
     create_payment_notification_for_order(
@@ -150,6 +181,7 @@ def handle_authorization(notification: Dict[str, Any], gateway_config: GatewayCo
         order_captured(payment.order, None, transaction.amount, payment)
 
 
+@transaction.atomic
 def handle_cancellation(notification: Dict[str, Any], _gateway_config: GatewayConfig):
     # https://docs.adyen.com/checkout/cancel#cancellation-notifciation
     payment = get_payment(notification.get("merchantReference"))
@@ -189,11 +221,14 @@ def handle_cancel_or_refund(
         handle_cancellation(notification, gateway_config)
 
 
+@transaction.atomic
 def handle_capture(notification: Dict[str, Any], _gateway_config: GatewayConfig):
     # https://docs.adyen.com/checkout/capture#capture-notification
     payment = get_payment(notification.get("merchantReference"))
     if not payment:
         return
+    checkout = get_checkout(payment)
+
     transaction_id = notification.get("pspReference")
     if payment.charge_status == ChargeStatus.FULLY_CHARGED:
         # the payment has already status captured.
@@ -206,6 +241,16 @@ def handle_capture(notification: Dict[str, Any], _gateway_config: GatewayConfig)
     gateway_postprocess(new_transaction, payment)
 
     reason = notification.get("reason", "-")
+
+    if not payment.order and new_transaction.is_success and checkout:
+        try:
+            order = create_order_from_checkout(checkout)
+        except Exception:
+            payment_refund_or_void(payment)
+            return
+        checkout.delete()
+        payment.order = order
+
     success_msg = f"Adyen: The capture {transaction_id} request was successful."
     failed_msg = f"Adyen: The capture {transaction_id} request failed. Reason: {reason}"
     create_payment_notification_for_order(
@@ -215,6 +260,7 @@ def handle_capture(notification: Dict[str, Any], _gateway_config: GatewayConfig)
         order_captured(payment.order, None, new_transaction.amount, payment)
 
 
+@transaction.atomic
 def handle_failed_capture(notification: Dict[str, Any], _gateway_config: GatewayConfig):
     # https://docs.adyen.com/checkout/capture#failed-capture
     payment = get_payment(notification.get("merchantReference"))
@@ -241,6 +287,7 @@ def handle_failed_capture(notification: Dict[str, Any], _gateway_config: Gateway
     create_payment_notification_for_order(payment, msg, None, True)
 
 
+@transaction.atomic
 def handle_pending(notification: Dict[str, Any], gateway_config: GatewayConfig):
     # https://docs.adyen.com/development-resources/webhooks/understand-notifications#
     # event-codes"
@@ -264,6 +311,7 @@ def handle_pending(notification: Dict[str, Any], gateway_config: GatewayConfig):
     )
 
 
+@transaction.atomic
 def handle_refund(notification: Dict[str, Any], _gateway_config: GatewayConfig):
     # https://docs.adyen.com/checkout/refund#refund-notification
     payment = get_payment(notification.get("merchantReference"))
@@ -296,6 +344,7 @@ def _get_kind(transaction: Optional[Transaction]) -> str:
     return TransactionKind.CAPTURE
 
 
+@transaction.atomic
 def handle_failed_refund(notification: Dict[str, Any], gateway_config: GatewayConfig):
     # https://docs.adyen.com/checkout/refund#failed-refund
     payment = get_payment(notification.get("merchantReference"))
@@ -359,6 +408,7 @@ def handle_failed_refund(notification: Dict[str, Any], gateway_config: GatewayCo
         gateway_postprocess(new_transaction, payment)
 
 
+@transaction.atomic
 def handle_reversed_refund(
     notification: Dict[str, Any], _gateway_config: GatewayConfig
 ):
@@ -395,6 +445,7 @@ def handle_refund_with_data(
     handle_refund(notification, gateway_config)
 
 
+@transaction.atomic
 def webhook_not_implemented(
     notification: Dict[str, Any], gateway_config: GatewayConfig
 ):
